Implement Marshal.GetIDispatchForObject on platforms with COM support
