+++ src/libraries/System.Runtime.InteropServices/tests/System/Runtime/InteropServices/Marshal/GetIDispatchForObjectTests.cs
@@ -13,9 +13,24 @@ namespace System.Runtime.InteropServices.Tests
     public partial class GetIDispatchForObjectTests
     {
         [Fact]
+        [PlatformSpecific(TestPlatforms.AnyUnix)]
         public void GetIDispatchForObject_NetCore_ThrowsPlatformNotSupportedException()
         {
             Assert.Throws<PlatformNotSupportedException>(() => Marshal.GetIDispatchForObject(null));
+        }
+
+        [Fact]
+        [PlatformSpecific(TestPlatforms.Windows)]
+        public void GetIDispatchForObject_NullObject_ThrowsArgumentNullException()
+        {
+            AssertExtensions.Throws<ArgumentNullException>("o", () => Marshal.GetIDispatchForObject(null));
+        }
+
+        [Fact]
+        [PlatformSpecific(TestPlatforms.Windows)]
+        public void GetIDispatchForObject_NonDispatchObject_ThrowsInvalidCastException()
+        {
+            Assert.Throws<InvalidCastException>(() => Marshal.GetIDispatchForObject(string.Empty));
         }
     }
 }
